MW-938 Get Customer Profile API tests have to be adjusted - added gatlin tests added min max checks

+++ src/main/java/com/icthh/xm/tmf/ms/customer/domain/properties/CustomerCharacteristics.java
@@ -25,9 +25,9 @@ public class CustomerCharacteristics {
         @JsonProperty("defaultValue")
         private String defaultValue;
         @JsonProperty("min")
-        private Long min;
+        private Integer min;
         @JsonProperty("max")
-        private Long max;
+        private Integer max;
         @JsonProperty("predefinedValues")
         private List<String> predefinedValues;
     }
